Extracted BuyButtonFactory and re-added buy button close to construction queue
